👌🏗 improve recorder architecture
* inline 'trackSegmentRenewal' in 'segmentCollection'
* move 'getSegmentContext' in 'segmentCollection'
* extract 'Segment' in its own file
* create workers and writers directly in 'segmentCollection'
