+++ packages/rum-core/src/index.ts
@@ -9,7 +9,7 @@ export {
   RumResourceEvent,
   RumLongTaskEvent,
 } from './rumEvent.types'
-export { CommonContext } from './rawRumEvent.types'
+export { ViewContext, CommonContext } from './rawRumEvent.types'
 export { startRum } from './boot/rum'
 export { LifeCycle, LifeCycleEventType } from './domain/lifeCycle'
 export { ParentContexts } from './domain/parentContexts'
